env_dump: get rid of most of the load-time link dependencies
The libcrypto is now loaded on demand and will hopefully use the version
already loaded by the binary we are tracing.
The xrandr lib was not currently necessary, so it got dropped.

The X11 codepath should only be called if the program already calls
a libX11/libxcb codepath, so there is nothing to do there.