env_dump: rework the internals to be thread safe and support dlsym
Some apps are using dlopen and dlsym for everything and we would fail to hook
anything in them. Fix this by keeping track of the symbols requested by the
clients using dlsym and, instead, return the hooked version of the symbol.

Instead of keeping track of all the hooks we have, query if env_dump.so
contains the symbol or not. To do so, we need to dlopen env_dump.so first
which requires reading the full path to it from the list of linked libs.

Once the hook has been executed, it needs to call the wanted function. Do so
by looking up the database of symbols. As some functions may be called
thousands of times per second, I also introduce a lookup-by-index function
which avoids the linear search.

This increases the cost of the lib a little but I still cannot measure it on
glxgears with 13k FPS.

All in all, easy right?