ezbench.py: support bisecting when there are build-failures
The idea is that every benchmark that we got asked to run on a commit
but could not be run because the commit does not compile, should be
moved to respectively before and after the commit that broke the build
and right after it.

This is sub-optimal in a bisecting scenario, but does not sound too
unreasonable. It also fixes a lot of issues, so let's go with this
for now.