core: make the profiles handle the repo
This commit sort of rewrites the entire world, but it changes the core
so as profiles can handle repos themselves. It allows having profiles
that do not have repos but also it allows caching deployments and
also calling the build system ourselves.

Most of the patch is just to make the already-existing profiles work
again!